Implement BookContainer
